Adding and Updating Files
IsAuthenticated middleware updated to verify the jwt if it is valid. I've terminated the functionalities of adding and removing categories and also i created a new functonality that lists all of the categories registers in database.

+++ backend/src/categories/services/categories.remove.service.ts
@@ -4,8 +4,9 @@ import { prisma } from '../../../prisma/prismaClient';
 
 // This functionality deletes a category register
 const removeCategory = async (data: category) => {
-    // Searching for category in database and returning the id
-    const categoryID = await prisma.category.findFirstOrThrow({
+
+    // Searching for category register in database and returning the id
+    const categoryID = await prisma.category.findFirst({
         where: {
             name: data.categoryName
         },
@@ -14,16 +15,29 @@ const removeCategory = async (data: category) => {
         }
     });
 
-    // Deleting category register in database
-    const result = await prisma.category.delete({
-        where: {
-            id: categoryID.id
-        }
-    });
+    // Checking the database operation result
+    if (!categoryID) {
 
-    // Returning the operation result
-    return result;
+        // In this case wasn't possible to find the category register in database
+        // Returning an error message
+        return "Non-Existent Category Register";
+
+    } else {
+
+        // In this case was possible to find the specific category
+        // Deleting category register in database
+        const result = await prisma.category.delete({
+            where: {
+                id: categoryID.id
+            }
+        });
+
+        // Returning the operation result
+        return result;
+
+    }
+
 }
 
 // Exporting Area
-export { removeCategory }
+export { removeCategory };
